SEL log Generator ID should Conform to IPMI Spec
In IMPI specification, the generator ID for SEL log is 2 bytes long.
Apart from the Generator ID, it contains LUN and channel information.
The fix is to change 1-Byte Generator ID to 2-byte generator ID and
include all missing information into it.

Tested
Following tests were executed:
1) Run ipmitool raw -l 1 0x4 0x2 0x4 0x8 0x1 0x6 0x1 0x5
This ensures that non-zero LUN information is included
in the SEL record.

2) Instrument the code to look like ME has created an event
and make sure that logs are saved properly.

For 1) and 2) logs are read from the WebUI to make sure that
they are being saved as expected.

+++ include/ipmi_to_redfish_hooks.hpp
@@ -26,7 +26,7 @@ bool checkRedfishHooks(uint16_t recordID, uint8_t recordType,
                        uint8_t sensorType, uint8_t sensorNum, uint8_t eventType,
                        uint8_t eventData1, uint8_t eventData2,
                        uint8_t eventData3);
-bool checkRedfishHooks(uint8_t generatorID, uint8_t evmRev, uint8_t sensorType,
+bool checkRedfishHooks(uint16_t generatorID, uint8_t evmRev, uint8_t sensorType,
                        uint8_t sensorNum, uint8_t eventType, uint8_t eventData1,
                        uint8_t eventData2, uint8_t eventData3);
 namespace redfish_hooks
